Move prof navbar to navbar.tsx
