+++ src/pages/Practice.tsx
@@ -3,16 +3,12 @@ import { H1, H2 } from '../components/atoms/typography'
 import { HomeBackground } from '../components/home/background'
 import { InputField } from '../components/atoms/input_field'
 import { NavUnlisted } from '../components/atoms/navLink'
-import { NavBarProf } from '../components/atoms/nav_bar_prof'
 
 // Page to display created components
 
 export const Practice = () => {
     return (
         <div>
-            <header>
-                <NavBarProf />
-            </header>
             <HomeBackground>
                 <H1>Welcome to the practice page</H1>
                 <MediumWhiteButton>
