+++ src/pages/ProfHomepage.tsx
@@ -5,12 +5,12 @@ import { InputField } from '../components/atoms/input_field'
 import { NavUnlisted } from '../components/atoms/navLink'
 import 'bootstrap/dist/css/bootstrap.css'
 import '../components/ProfHomepage/ProfHomepage.css'
-import { NavBarAdmin } from '../components/navbar'
+import { NavBarAdmin, NavBarProf } from '../components/navbar'
 
 export const ProfHomepage = () => {
     return (
         <div>
-            <NavBarAdmin />
+            <NavBarProf />
             <div className='cen'>
                 <div className='row'>
                     <div className='col'>
